refactor: move user schema to core
