+++ packages/better-auth/src/types/models.ts
@@ -2,7 +2,6 @@ import type { BetterAuthOptions } from "./options";
 import type {
 	accountSchema,
 	sessionSchema,
-	userSchema,
 	verificationSchema,
 } from "../db/schema";
 import type { Auth } from "../auth";
@@ -10,6 +9,7 @@ import type { InferFieldsFromOptions, InferFieldsFromPlugins } from "../db";
 import type { StripEmptyObjects, UnionToIntersection } from "./helper";
 import type { BetterAuthPlugin } from "./plugins";
 import type * as z from "zod";
+import type { User } from "@better-auth/core/db";
 
 export type Models =
 	| "user"
@@ -89,7 +89,7 @@ interface RateLimit {
 	lastRequest: number;
 }
 
-export type User = z.infer<typeof userSchema>;
+export type { User } from "@better-auth/core/db";
 export type Account = z.infer<typeof accountSchema>;
 export type Session = z.infer<typeof sessionSchema>;
 export type Verification = z.infer<typeof verificationSchema>;
